+++ src/blueprints/auth.py
@@ -1,7 +1,7 @@
-
+from datetime import datetime
 from flask import render_template, request, redirect, url_for,  Blueprint, flash
 
-from flask_login import login_user, logout_user
+from flask_login import login_user, logout_user, current_user, login_required
 
 from src.config import app, db, bcrypt, login_manager
 from src.models import User, log_action, Settings
@@ -19,9 +19,11 @@ def register():
         # Check if this is the first user
         if User.query.count() == 0:
             user_level = 'admin'  # Make the first user an admin
+            password_changed = False
         else:
             user_level = 'customer'  # Default to 'customer' for all other users
-        user = User(fname=fname, lname=lname, username=username, email=email, password=password, user_level=user_level)
+            password_changed = True
+        user = User(fname=fname, lname=lname, username=username, email=email, password=password, user_level=user_level, password_changed=password_changed)
         db.session.add(user)
         db.session.commit()
         return redirect(url_for('auth.login'))
@@ -37,11 +39,45 @@ def login():
         if user and bcrypt.check_password_hash(user.password, password):
             login_user(user)
             log_action(user.id, user.username, 'Login', f'User {user.username} logged in.')
+
+            # Check if the admin needs to change their password
+            if user.user_level == 'admin' and not user.password_changed:
+                flash('You need to change your password first to continue.', 'warning')
+                return redirect(url_for('auth.change_password'))
+            
             flash(f'Login successful as {user.username}', 'success')
             return redirect(url_for('admin.dashboard'))
         else:
             flash('Login failed. Please check your credentials.', 'danger')
     return render_template('login.html', settings=system_settings)
+
+@auth_bp.route("/change-password", methods=['GET', 'POST'])
+@login_required
+def change_password():
+    system_settings = Settings.query.first()
+    if request.method == 'POST':
+        current_password = request.form.get('current_password')
+        new_password = request.form.get('new_password')
+        user = User.query.get(current_user.id)
+
+        if user and bcrypt.check_password_hash(user.password, current_password):
+            user.password = bcrypt.generate_password_hash(new_password).decode('utf-8')
+            user.password_changed = True  # Set the flag to True
+            user.last_password_change=datetime.utcnow()
+            db.session.commit()
+            flash('Password updated successfully!', 'success')
+
+            return redirect(url_for('auth.change_password'))
+        else:
+            flash('Current password is incorrect.', 'danger')
+
+    # if current user password_changed false, flash a message
+    if not current_user.password_changed:
+        flash('You need to change your password first to continue.', 'warning')
+        return render_template('change_password.html', settings=system_settings)
+
+    return render_template('change_password.html', settings=system_settings)
+        
 
 @auth_bp.route("/logout")
 def logout():